refactor: extract analytics routes
